refactor(tests): separate codex history integration
